Rewrite of css inliner functionality to be wrapped in MailDeliverService
