+++ tachikoma-it/build.gradle.kts
@@ -22,4 +22,5 @@ dependencies {
     testImplementation("org.apache.logging.log4j:log4j-slf4j-impl:$log4j2Version")
     testImplementation("com.linecorp.armeria:armeria:$armeriaVersion")
     testImplementation("com.squareup.okhttp3:okhttp:4.4.0")
+    testImplementation("org.jsoup:jsoup:1.12.2")
 }
