+++ tachikoma-it/src/test/resources/wrapAndPackBody/parseHTML/simple/expectedNoInlining.html
@@ -1,0 +1,16 @@
+<html>
+<head>
+  <style type="text/css">
+    body {
+      margin: 0;
+    }
+    .jiffy {
+      font-size: large;
+      color: #f47a39;
+    }
+  </style>
+</head>
+<body>
+<h1 class="jiffy">Hello World</h1>
+</body>
+</html>
